requirements.txt
pdfplumber
pymupdf
requests
pandas
pydantic
deep-translator
camelot-py
ghostscript